backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "alembic>=1.13.0",
    "anthropic>=0.40.0",
    "black==24.1.1",
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "email-validator>=2.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "fastapi>=0.115.0",
    "fastmcp==0.2.0",
    "flake8==7.0.0",
    "google-auth>=2.23.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    "httpx>=0.27.0",
    "langchain==0.3.27",
    "langchain-anthropic==0.3.6",
    "langchain-aws==0.2.35",
    "langchain-community==0.3.27",
    "langchain-core>=0.3.72,<1.0.0",
    "langchain-openai==0.3.35",
    "langgraph==0.5.3",
    "loguru==0.7.2",
    "mcp==1.0.0",
    "mypy==1.8.0",
    "neo4j==5.16.0",
    "openai>=1.54.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.7.0",
    "pymilvus==2.3.6",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "tenacity==8.2.3",
    "tiktoken>=0.7.0",
    "uvicorn[standard]>=0.32.0",
]
